refactor: correct wrong spelling
