+++ src/components/Deposit/EditConfirmModal.tsx
@@ -58,7 +58,7 @@ const EditConfirmModal = (props: Props) => {
           <ModalCloseButton />
           <ModalBody textAlign="center" marginTop="60px">
             <NumberInput defaultValue={currentDeposit} precision={2} step={0.01} max={1} min={0}>
-              <NumberInputField ref={ref} data-testid="numberInput" />
+              <NumberInputField ref={ref} data-testid="numberInput" autoFocus />
               <NumberInputStepper>
                 <NumberIncrementStepper />
                 <NumberDecrementStepper />
